text-to-3D/requirements_threestudio.txt
absl-py==2.1.0
accelerate==0.29.2
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrdict==2.0.1
attrs==23.2.0
av==12.0.0
awscli==1.32.82
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
bitsandbytes==0.38.1
bleach==6.1.0
blessed==1.20.0
blinker==1.7.0
boto3==1.34.82
botocore==1.34.82
Brotli==1.0.9
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==2.0.4
click==8.1.7
clip==1.0
colorama==0.4.4
colorlog==6.8.2
comet-ml==3.40.0
comm==0.2.2
ConfigArgParse==1.7
configobj==5.0.8
contourpy==1.1.1
controlnet_aux==0.0.7
croniter==1.3.15
cryptography==42.0.5
cycler==0.12.1
dash==2.16.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dateutils==0.6.12
debugpy==1.8.1
decorator==5.1.1
deepdiff==7.0.1
defusedxml==0.7.1
descartes==1.1.0
diffusers==0.19.3
dill==0.3.8
docker-pycreds==0.4.0
docstring_parser==0.16
docutils==0.16
dulwich==0.21.7
editor==1.6.6
einops==0.7.0
embreex==2.17.7.post4
envlight==0.1.0
eval-type-backport==0.1.3
everett==3.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.1.17
fastjsonschema==2.19.1
ffmpy==0.3.2
filelock==3.13.1
fire==0.6.0
Flask==3.0.3
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.12.2
ftfy==6.2.0
gdown==5.1.0
gitdb==4.0.11
GitPython==3.1.43
gmpy2==2.1.2
google-auth==2.29.0
google-auth-oauthlib==1.0.0
gradio==4.11.0
gradio_client==0.7.3
grpcio==1.62.1
gsplat==0.1.10
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.4
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib_metadata==7.1.0
importlib_resources==6.4.0
inquirer==3.2.4
ipykernel==6.29.4
ipython==8.12.3
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.1.2
jaxtyping==0.2.19
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.0
json5==0.9.24
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter_client==8.6.1
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.26.0
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
kornia==0.7.2
kornia_rs==0.1.3
lazy_loader==0.4
libigl==2.5.1
lightning==2.0.0
lightning-cloud==0.5.65
lightning-utilities==0.11.2
lxml==5.2.1
mapbox-earcut==1.0.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.5
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapy==1.2.0
mistune==3.0.2
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
msgpack==1.0.8
msgpack-numpy==0.4.8
multidict==6.0.5
multiprocess==0.70.16
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nerfacc==0.5.2
nerfstudio==1.0.3
nest-asyncio==1.6.0
networkx==3.1
ninja==1.11.1.1
nodeenv==1.8.0
notebook==7.1.2
notebook_shim==0.2.4
numpy==1.24.3
nuscenes-devkit==1.1.9
nvdiffrast==0.3.1
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.19.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
open3d==0.18.0
opencv-python==4.8.0.76
ordered-set==4.1.0
orjson==3.10.0
overrides==7.7.0
packaging==24.0
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
pathos==0.3.2
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pip==24.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.0
plotly==5.20.0
pox==0.3.4
ppft==1.7.6.8
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.7
pycollada==0.8
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
PyJWT==2.8.0
pyliblzfse==0.4.1
PyMCubes==0.1.4
pymeshlab==2023.12.post1
pyngrok==7.1.6
pyparsing==3.1.2
pyquaternion==0.9.9
pysdf==0.1.9
PySocks==1.7.1
python-box==6.1.0
python-dateutil==2.9.0.post0
python-engineio==4.9.0
python-json-logger==2.0.7
python-multipart==0.0.9
python-socketio==5.11.2
pytorch-lightning==2.2.1
pytorch-msssim==1.0.0
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
rawpy==0.19.1
readchar==4.0.6
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.7.2
Rtree==1.2.0
runs==1.2.2
s3transfer==0.10.1
safetensors==0.4.2
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.2.0
sentry-sdk==1.45.0
setproctitle==1.3.3
setuptools==69.2.0
shapely==2.0.3
shellingham==1.5.4
shtab==1.7.1
simple-websocket==1.0.0
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
splines==0.3.0
stack-data==0.6.3
starlette==0.37.2
starsessions==1.3.0
svg.path==6.3
sympy==1.12
taming-transformers-rom1504==0.0.6
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.4.0
tifffile==2023.7.10
timm==0.6.7
tinycss2==1.2.1
tinycudann==1.7
tokenizers==0.13.3
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.1.1+cu118
torch-fidelity==0.3.0
torchaudio==2.1.1
torchmetrics==1.3.2
torchvision==0.16.1+cu118
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
transformers==4.28.1
trimesh==4.3.0
triton==2.1.0
typeguard==4.2.1
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tyro==0.8.3
tzdata==2024.1
uri-template==1.3.0
urllib3==1.26.18
uvicorn==0.29.0
vhacdx==0.0.6
viser==0.1.27
wandb==0.16.6
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.3.3
websockets==11.0.3
Werkzeug==3.0.2
wheel==0.41.2
widgetsnbextension==4.0.10
wrapt==1.16.0
wsproto==1.2.0
wurlitzer==3.0.3
xatlas==0.0.9
xformers==0.0.23
xmod==1.8.1
xxhash==3.4.1
yarl==1.9.4
yourdfpy==0.0.56
zipp==3.18.1
